requirements.txt
python-telegram-bot
python-dotenv
openai
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-openai 